Fallback to GitHub if jenkins job build info does not contain repos/commits
A job that has just started for the first time (and other times too if it is
running on a worker which has not executed before) will not provide
any repository/commits information when we request its information.

In that case we fallback to parsing the configuration and requesting
that information from GitHub.